requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.1
alembic==1.11.2
altair==5.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bbaug==0.4.2
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.3
cvzone==1.5.6
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
ecdsa==0.18.0
exceptiongroup==1.1.2
extra-streamlit-components-better-cookie-manager==0.0.10
fastapi==0.100.0
filelock==3.12.2
filterpy==1.4.5
fonttools==4.40.0
frozenlist==1.4.0
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.20.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.56.0
h11==0.14.0
hydra-core==1.3.2
idna==2.10
imageio==2.31.1
imgaug==0.4.0
importlib==1.0.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.23.3
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lap==0.4.0
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.0
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.74
packaging @ file:///C:/b/abs_ed_kb9w6g4/croot/packaging_1678965418855/work
pandas==2.0.3
pascal-voc-writer==0.1.4
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
pluggy==1.2.0
ply==3.11
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
psycopg2==2.9.6
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycparser==2.21
pydantic==1.10.10
pydantic_core==2.4.0
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip @ file:///C:/Windows/Temp/abs_d7gmd2jg8i/croots/recipe/pyqt-split_1659273064801/work/pyqt_sip
PyQt6==6.5.1
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.0
qudida==0.0.4
referencing==0.30.0
requests==2.23.0
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.9.2
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.3.0
scipy==1.11.0
seaborn==0.12.2
sentry-sdk==1.26.0
sh==1.14.1
shapely==2.0.1
simplejson==3.19.1
sip @ file:///C:/Windows/Temp/abs_b8fxd17m2u/croots/recipe/sip_1659012372737/work
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.31.1
streamlit==1.25.0
streamlit-aggrid==0.3.4
streamlit-autorefresh==1.0.1
streamlit-image-annotation==0.3.1
streamlit-option-menu==0.3.6
streamlit-tags==1.2.8
sympy==1.12
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==2.0.1
toolz==0.12.0
torch==2.1.0.dev20230624+cu121
torchaudio==2.1.0.dev20230625+cu121
torchvision==0.16.0.dev20230625+cu121
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.100
urllib3==1.25.11
uvicorn==0.22.0
validators==0.20.0
watchdog==3.0.0
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
